refactor: move tests related to listener denials
